Add simple C interface to libcadet
Adds a simple C interface to libcadet. The parameter provider interface
is realized by callback functions. A simulation can be performed and
the solution at the outlet of the unit operations can be queried.

Also includes a test application that dynamically loads the library at
runtime (only Windows) and executes a simulation specified in JSON
format internally.

This commit provides the first steps regarding #16 and, thus, #12.